refactor chat node helpers
